* Compiler warning fix
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@825 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ spells.cpp
@@ -846,7 +846,7 @@ int32_t Spell::getManaCost(const Player* player) const
 		return mana;
 
 	if(player && manaPercent != 0)
-		return std::floor((player->getMaxMana() * manaPercent) / 100);
+		return (int32_t)std::floor(double(player->getMaxMana() * manaPercent) / 100);
 
 	return 0;
 }
